+++ src/backend/renderer/mod.rs
@@ -14,7 +14,7 @@ use crate::utils::{Buffer, Physical, Point, Rectangle, Size};
 
 #[cfg(feature = "wayland_frontend")]
 use crate::wayland::compositor::SurfaceData;
-use cgmath::{prelude::*, Matrix3, Vector2, Vector3};
+use cgmath::Matrix3;
 #[cfg(feature = "wayland_frontend")]
 use wayland_server::protocol::{wl_buffer, wl_shm};
 
@@ -51,6 +51,15 @@ pub enum Transform {
     Flipped180,
     /// Plane is flipped vertically and rotated by 270 degrees
     Flipped270,
+}
+
+#[derive(Debug, PartialEq, Eq, Hash, Clone, Copy)]
+/// Texture filtering methods
+pub enum TextureFilter {
+    /// Returns the value of the texture element that is nearest (in Manhattan distance) to the center of the pixel being textured.
+    Linear,
+    /// Returns the weighted average of the four texture elements that are closest to the center of the pixel being textured.
+    Nearest,
 }
 
 impl Transform {
@@ -163,25 +172,17 @@ pub trait Frame {
     /// This operation is only valid in between a `begin` and `finish`-call.
     /// If called outside this operation may error-out, do nothing or modify future rendering results in any way.
     fn clear(&mut self, color: [f32; 4]) -> Result<(), Self::Error>;
-    /// Render a texture to the current target using given projection matrix and alpha.
-    /// The given vertices are used to source the texture. This is mostly useful for cropping the texture.
-    fn render_texture(
-        &mut self,
-        texture: &Self::TextureId,
-        matrix: Matrix3<f32>,
-        tex_coords: [Vector2<f32>; 4],
-        alpha: f32,
-    ) -> Result<(), Self::Error>;
 
     /// Render a texture to the current target as a flat 2d-plane at a given
     /// position and applying the given transformation with the given alpha value.
+    /// (Meaning `src_transform` should match the orientation of surface being rendered).
     fn render_texture_at(
         &mut self,
         texture: &Self::TextureId,
         pos: Point<f64, Physical>,
         texture_scale: i32,
         output_scale: f64,
-        transform: Transform,
+        src_transform: Transform,
         alpha: f32,
     ) -> Result<(), Self::Error> {
         self.render_texture_from_to(
@@ -195,55 +196,22 @@ pub trait Frame {
                     .to_f64()
                     .to_physical(output_scale),
             ),
-            transform,
+            src_transform,
             alpha,
         )
     }
 
-    /// Render part of a texture as given by src to the current target into the rectangle described by dest
-    /// as a flat 2d-plane after applying the given transformations.
+    /// Render part of a texture as given by src to the current target into the rectangle described by dst
+    /// as a flat 2d-plane after applying the inverse of the given transformation.
+    /// (Meaning `src_transform` should match the orientation of surface being rendered).
     fn render_texture_from_to(
         &mut self,
         texture: &Self::TextureId,
         src: Rectangle<i32, Buffer>,
-        dest: Rectangle<f64, Physical>,
-        transform: Transform,
+        dst: Rectangle<f64, Physical>,
+        src_transform: Transform,
         alpha: f32,
-    ) -> Result<(), Self::Error> {
-        let mut mat = Matrix3::<f32>::identity();
-
-        // position and scale
-        mat = mat * Matrix3::from_translation(Vector2::new(dest.loc.x as f32, dest.loc.y as f32));
-        mat = mat * Matrix3::from_nonuniform_scale(dest.size.w as f32, dest.size.h as f32);
-
-        //apply surface transformation
-        mat = mat * Matrix3::from_translation(Vector2::new(0.5, 0.5));
-        if transform == Transform::Normal {
-            assert_eq!(mat, mat * transform.invert().matrix());
-            assert_eq!(transform.matrix(), Matrix3::<f32>::identity());
-        }
-        mat = mat * transform.invert().matrix();
-        mat = mat * Matrix3::from_translation(Vector2::new(-0.5, -0.5));
-
-        // this matrix should be regular, we can expect invert to succeed
-        let tex_size = texture.size();
-        let texture_mat = Matrix3::from_nonuniform_scale(tex_size.w as f32, tex_size.h as f32)
-            .invert()
-            .unwrap();
-        let verts = [
-            (texture_mat * Vector3::new((src.loc.x + src.size.w) as f32, src.loc.y as f32, 0.0)).truncate(), // top-right
-            (texture_mat * Vector3::new(src.loc.x as f32, src.loc.y as f32, 0.0)).truncate(), // top-left
-            (texture_mat
-                * Vector3::new(
-                    (src.loc.x + src.size.w) as f32,
-                    (src.loc.y + src.size.h) as f32,
-                    0.0,
-                ))
-            .truncate(), // bottom-right
-            (texture_mat * Vector3::new(src.loc.x as f32, (src.loc.y + src.size.h) as f32, 0.0)).truncate(), // bottom-left
-        ];
-        self.render_texture(texture, mat, verts, alpha)
-    }
+    ) -> Result<(), Self::Error>;
 }
 
 /// Abstraction of commonly used rendering operations for compositors.
@@ -255,6 +223,11 @@ pub trait Renderer {
     /// Type representing a currently in-progress frame during the [`Renderer::render`]-call
     type Frame: Frame<Error = Self::Error, TextureId = Self::TextureId>;
 
+    /// Set the filter method to be used when rendering a texture into a smaller area than its size
+    fn downscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>;
+    /// Set the filter method to be used when rendering a texture into a larger area than its size
+    fn upscale_filter(&mut self, filter: TextureFilter) -> Result<(), Self::Error>;
+
     /// Initialize a rendering context on the current rendering target with given dimensions and transformation.
     ///
     /// This function *may* error, if:
@@ -265,7 +238,7 @@ pub trait Renderer {
     fn render<F, R>(
         &mut self,
         size: Size<i32, Physical>,
-        transform: Transform,
+        dst_transform: Transform,
         rendering: F,
     ) -> Result<R, Self::Error>
     where